backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.3.2
pymongo>=4.6.3
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
pydantic>=2.7.0
pydantic-settings>=2.2.1
pydantic[email]>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
gitpython>=3.1.43
